Add platform agnostic implementation of Path::from_bytes()
The implementation of the Path::from_bytes() method is the same on the
unix platform and on the windows platform. This is a maintenance burden;
if any changes are needed to be made to this method, then it will more
than likely have to be applied on each platform.

This change unifies the Path::from_bytes() method so that it is
implemented only once making it easier to maintain.