Rename os_str_to_bytes() -> os_str_as_bytes()
The os_str_to_bytes() method implies that an allocation happens in order
to create a byte array; in rust, there is a convention where a function
named to_* implies that an allocation happens. However,
os_str_to_bytes() does not allocate any memory and instead only returns
a reference to already existing memory.

This change renames the function to os_str_as_bytes() so that it is in
keeping with rust naming convention; naming a function as_* generally
implies that no allocation occurs.