Move Path::as_bytes() method into platform agnostic Path impl
The implementation of the Path type only defines a single method --
as_bytes() -- on each platform. This is a maintenance burden and also
makes it harder to understand the code since these implementations are
tucked away in platform specific modules.

This change moves the Path::as_bytes() method so that it is implemented
in the same platform agnostic module as the Path type is defined. While
still needing to be implemented differently based on platform, having
only a single module where the method is implemented makes it it much
easier to understand and maintain.