Move From and AsRef trait impl on Path into crate::path module
The implementation of `From<&Path> for Vec<u8>` and `AsRef<[u8]> for
Path` is the same for each platform, however each platform module
duplicates the implementation. This is a maintenance burden as if there
are any changes to these 2 implementations, the changes will need to be
applied twice (ie once for unix and another for windows).

This change moves these implementations out of the platform specific
modules and into the crate::path module. This means that now there is
only a single implementation making it much easier to maintain.