Move PathBuf impl into crate::path module
The PathBuf type is implemented in each platform specific module, once
for the unix platform, and again for the windows platform. However,
these implementations are largely the same. This means it is harder to
maintain going forward as any changes to the PathBuf struct of one
platform likely means the change will have to be re-interpreted and
applied to the PathBuf struct of the other platform.

This change moves the implementation of the PathBuf type into the
crate::path module and out of the platform specific modules. While there
is still some need for platform specific code, it is much easier to
maintain and understand as the complete definition of the PathBuf type
is now in a single place.