Add platform neutral impl of a windows file path iterator
Since the implementation of a iterator over a windows file path is now
largely the same for the windows and unix platforms, this constitutes a
lot of duplicated code. This makes it harder to maintain as any changes
to the implementation on one platform will have to be made again on the
other platform.

This change unifies the platform specific implementations into a single
implementation of a windows file path iterator. This makes maintaining
the implementation of the iterator much easier as there is only one
module where the iterator exists.