Update unit tests to remove platform specific tests
The unit and integration tests are both split by platform; the unix
platform has its own tests and so does the windows platform. However,
these tests are the same now due to unifying the path types into a
single platform agnostic module. This means that there is now a lot of
needlessly duplicated test code.

This change removes the platform split in the tests and, in doing so,
removes duplicated test code. As a result, maintaining the test suite is
thus made much easier.