Rename Path and PathBuf -> PlatformPath and PlatformPathBuf
The Path and PathBuf types seem more like trait names than type names.
As such, future commits will implement these names as traits. However,
this clashes with the current concrete types.

This change renames the Path and PathBuf types to PlatformPath and
PlatformPathBuf. This frees up the Path and PathBuf names so that they
can be used as the names of traits.