Add path traits
Some methods of the platform path types should be available on all path
types. However, as these methods are inherent methods, there is no way
to specify these path types when using generics.

This change moves the as_bytes() and as_os_str() methods inherent to the
PlatformPath and PlatformPathBuf concrete types into a new Path trait.
This makes it possible to specify a general path type when using
generics.