path: gather all traits in one place
The path module defines traits at the beginning of the file and at the
end. This makes it easier to not realize that there are more traits in
the file when reading top to bottom.

This change moves the MemoryPath traits from the bottom of the path
module's file to near the top with the other traits. As a result, all
trait definitions are in the same section of the file which makes it
easier to read and understand and maintain.