Rename Path -> SystemSeq
The Path trait was named as the original intent was for the trait to
represent a path. However, as the design of the crate was iterated, this
responsibility was pushed onto the MemoryPath trait. The only thing
remaining for the now inaccurately named Path trait was specifying an
interface to access system strings as either a UTF-8 ish byte string or
as an os string.

This change renames the Path trait to SystemSeq so that the trait more
accurately communicates what it provides.