Rename PathBuf -> SystemSeqBuf
The PathBuf trait was named as the original intent was for the trait to
represent a path buffer (ie a path type that could be manipulated).
However, as the design of the crate was iterated, this responsibility
was pushed onto the MemoryPathBuf trait. This resulted in the PathBuf
now being inaccurately named as it no longer represents a path type.

This change renames the PathBuf trait to SystemSeqBuf so that the trait
more accurately communicates what it provides.