Add new method: MemoryPath::parts()
Iterating over the path components of a path is provided by the
MemoryPath::iter() method. However, the items returned by the iterator
require matching against the type of path. In some cases where this is
not needed and it is only wanted to get the components as strings, then
extra work must be done to convert the returned types into strings.
Also, as the item returned by the iterator is a Result type, any errors
must also be handled which is added complexity.

This change adds a new method MemoryPath::parts(). This method when
called returns an iterator that returns path components as OsString
types. If an error occurs when iterating over a path, then the error
marks the end of iteration; the error is just swallowed in this case and
is not propagated to the caller of the parts() method.

With the MemoryPath::parts() method, there now exists a quick and simple
way to just get path components as strings without having to worry about
error handling or string conversion of component types.