Add new trait: MemoryPathPartsExt
For some platforms, implementing an iterator for the MemoryPathParts
type (which is meant to be an iterator that converts path components
into an OsString type) requires both accessing the path's iterator over
component parts as well as accessing and storing an item. However, as
both of these are private attributes, and getter methods are
undocumented (due to the intention that these getter methods should only
be used when implementing an iterator for MemoryPathParts), there is no
way to extend MemoryPathParts to other platforms.

This change moves the getter methods into a new trait,
MemoryPathPartsExt. This makes it explicit when needing to use these
getter methods, as well as able to be documented in the future that the
trait is only intended to be used when extending the MemoryPathParts
type in some way.