Rename MemoryPath -> Path
The names of the MemoryPath and MemoryPathBuf traits and the related
traits and types MemoryPathParts struct and MemoryPathPartsExt trait,
all are confusing names as they have practically nothing to do with
memory beyond being held and manipulated solely in memory (which isn't
something significant enough to explicitly point out).

This change renames these traits and types to be based off what they
actually are: general paths. This means that the names of the main
traits are Path and PathBuf, and the related traits and types are now
PathParts and PathPartsExt.