Return reference only when creating new Path
UnixPath and WindowsPath return a concrete object when their respective
static new() methods are called. However, as types implementing the Path
trait are meant to be seen similarly as the stdlib's Path and str types
ie views into owned data, then returning concrete objects does not
properly communicate this. This can be a cognitive disconnect when
compared to the convention used by the rust language and its stdlib ie
that views into owned data are explicitly references.

This change makes it so that calling UnixPath::new() or
WindowsPath::new() returns a reference instead of an owned type.

Please note that this adds unsafe blocks, however these blocks are
actually safe since only immutable references are being returned by the
new() methods.