Add new path trait: PathIterator
Currently, it is not possible to easily create new types that implement
the Path trait with different iterators. There currently can only be a
single path iterator specified for any given concrete Path trait
implementation; there cannot be multiple different types of Path trait
implementations where the only difference is the type of iterator being
used. This makes it harder to develop different ways of parsing paths as
there can only be a single way to do so at any given time.

In order to make it possible for there to be multiple concrete Path
trait implementations with different iterators, a new trait is added:
`PathIterator`.

PathIterator can only be implemented on types that already implement the
Iterator trait. In addition, PathIterator specifies a standard way that
the type can be instantiated. This is important as in future commits,
that Path trait wil be changed so that it uses a generic type instead of
an associated type, enabling multiple implementations of the Path trait
to exist for a single type where the difference will be in the iterator
used. This is important in the near future for benchmarking different
implementations of parsing paths (eg hand-coded parser vs using a more
general parsing library), and also to ensure the orthogonality of path
parsing with respect to the overall behaviour of a Path like type.