Add new trait: AsSystemStr
The Path trait currently requires that the iter() method be manually
implemented. However, as future commits will be adding more methods to
the Path trait, it is desirable to have as few methods as possible
needing to be manually implemented; not only will this make the
ergonomics of implementing new Path types easier, it will also reduce
fragility and duplication of effort.

In order to provide a default implementation of the Path trait's
iter() method, the type will need to somehow produce a &SystemStr in
to feed into the PathIterator::new() method. However, while the Deref ->
&SystemStr exists, it only is implemented for the types that directly
implement the Path trait; accessing &SystemStr from the PathBuf types,
while possible, is a bit unwieldy.

This change adds a new trait `AsSystemStr` that provides a common way to
access a path type's underlying &SystemStr type.

Please note that this change only defines the trait; future commits will
implement this trait for the various Path and PathBuf types.