Implement AsSystemStr for all path types
The Path trait currently requires that the iter() method be manually
implemented. However, as future commits will be adding more methods to
the Path trait, it is desirable to have as few methods as possible
needing to be manually implemented; not only will this make the
ergonomics of implementing new Path types easier, it will also reduce
fragility and duplication of effort.

In order to provide a default implementation of the Path trait's iter()
method, the type will need to somehow produce a &SystemStr in to feed
into the PathIterator::new() method. However, while the Deref ->
&SystemStr exists, it only is implemented for the types that directly
implement the Path trait; accessing &SystemStr from the PathBuf types,
while possible, is a bit unwieldy. A standard way to provide the path
types' underlying &SystemStr needs to be implemented.

This change implements the AsSystemStr trait that was added in commit
18f3f39e34ef ("Add new trait: AsSystemStr") for all Path and PathBuf
types; this trait provides a method that returns a &SystemStr. Future
commits can then use the implementation of this trait on the path types
to implement a default implementation of the Path trait's iter() method.

The consequence of this, is any type wanting to successfully implement
the Path type will need to first create an iterator that implements the
PathIterator trait, and the type implementing the Path trait will also
need to implement the AsSystemStr trait.