Add default impl for Path::iter()
Currently, it is not possible to easily create new types that implement
the Path trait with different iterators. There currently can only be a
single path iterator specified for any given concrete Path trait
implementation; there cannot be multiple different types of Path trait
implementations where the only difference is the type of iterator being
used. This makes it harder to develop different ways of parsing paths as
there can only be a single way to do so at any given time.

This change adds a default implementation for the Path trait's iter()
method. Implementations of the Path trait have also been updated to
ensure unit tests still all pass. Now, the minimum needed to implement a
new implementation of the Path type are the following:

* A concrete type implementing the AsSystemStr trait

* A concrete type implementing both the Iterator and PathIterator traits

* An Iterator implementation for the PathParts concrete type

This change depends on:

commit c75c25c759da ("Add new path trait: PathIterator")

commit 14099911298d ("Implement PathIterator trait for path iterator
types")

commit 18f3f39e34ef ("Add new trait: AsSystemStr")

commit 9e840d61a3d1 ("Implement AsSystemStr for all path types")