Move windows path iterator -> crate::windows::iter::manual
Currently, types that implement the Path trait can only use a single
iterator type to iterate the components of a path. It should be
possible, however, to implement different types of iterators where each
iterator could be a different implementation. Currently, this can only
be hard coded and always included.

The solution to this is to put different path iterator implementations
behind feature gates which could then be switched on or off.

This change moves the manual implementation of an iterator for
windows paths to a new module crate::windows::iter::manual. This new
module is gated behind the `manual-iter` feature which was added in
commit 62fa2b91ee159 ("Add new cargo feature: manual-iter").