Add new cargo feature: parser-iter
Currently, types that implement the Path trait can only use a single
iterator type to iterate the components of a path. It should be
possible, however, to implement different types of iterators where each
iterator could be a different implementation. Currently, this can only
be hard coded and always included.

The solution to this is to add feature gates to the project's Cargo.toml
file which each different type of iterator could then be included or
not.

This change adds a feature gate for an implementation of path iterators
using the `combine` parsing library.