fix: parser-iter feature failing windows invalid filename test
When using the parser-iter feature, the
test::windows::iter::invalid_filename test fails. This is because the
behaviour of the windows path parser implemented under the parser-iter
feature gate for dealing with invalid filenames is different than the
path parser implemented under the manual-iter feature gate.
Specifically, the issue is when parsing a path component, the parser
under the parser-iter feature gate does not take the case of the
component into consideration; it will only consider a component with an
invalid filename to be invalid if the case of the component is all
uppercase. However, windows path components are case insensitive, so the
expected behaviour is to always consider an invalid path component
invalid regardless of the case of the component.

For example, a path component named "NUL" is invalid as that is the name
of a windows device. This also means that variations such as "NUL.ext"
and "nul.ext" are also to be considered invalid. The parser-iter parser,
however, will only consider the "NUL" and "NUL.ext" cases as invalid in
this example, and accept "nul.ext" as valid.

This change ensures that the parser-iter parser does a case-insensitive
check when determining if a file component is invalid.