fix: parser-iter feature failing windows verbatim disk test
When using the parser-iter feature, the
test::windows::iter::verbatim_disk test fails. This is because the
behaviour of the windows path parser implemented under the parser-iter
feature gate for parsing verbatim disk windows prefixes is different
than the path parser implemented under the manual-iter feature gate.
Specifically, the issue is that the manual-iter parser creates a
Component::Prefix variant that holds the prefix "//?/C:\", whereas the
parser-iter parser creates a similar Component::Prefix variant holding
"//?/C:" (notice the missing trailing separator character).

The cause of this issue is due to the parser-iter parser not actually
consuming the trailing separator character in order to be able to parse
it as the path's root. This results in the trailing separator character
being not included in the parsed prefix.

To ensure that the behaviour of the manual-iter and parser-iter parser
implementations are the same, this commit ensures that the trailing
separator character is included in the parsed prefix returned by the
windows path iterator for verbatim disk prefixes while still being able
to parse the path's root.