Move ascii_uppercase function to crate::common::string module
When running tests against the parser-iter feature, a lot of warnings
are generated advising of code in the crate::windows::match_prefix
module not being used. While the solution is to put the match_prefix
module behind the manual-iter feature gate, this cannot be done as the
ascii_uppercase function defined in the match_prefix module is needed by
the parser-iter feature.

This change moves the ascii_uppercase function out of the match_prefix
module and into the crate::common::string module. As a result, this
function can now be used regardless of enabled feature gates.