Put windows::match_prefix module behind manual-iter feature gate
When running tests against the parser-iter feature, a lot of warnings
are generated advising of code in the crate::windows::match_prefix
module not being used.

This change puts the match_prefix module behind the manual-iter feature
gate so that the module is only compiled if the manual-iter feature is
enabled.