fix: windows::parser::file_name never returns None case
The windows::parser::file_name parser function parses windows file name
path components. If it is given an empty u8 array, then it is expected
that the function returns a None value. What is actually returned,
though, is an array containing 2 empty arrays. This is wasteful and
confusing, and also renders the conditional that returns the None value
as unreachable and/or dead code.

This change updates the file_name parser function so that it returns
None when parsing an empty u8 array. Included in this change is a unit
test validating this behaviour.