Remove parser-iter from list of default features
While the parser-iter feature was in development, it was activated by
default along with the manual-iter feature. However, now that the
feature is done, there is no need for it to be a default feature; using
a parser library to iterate over a path has always been intended to be
optional until it can be benchmarked against the manual implementation
of path iteration.

This change removes the parser-iter feature from the list of default
features. With this change, the parser-iter feature must be explicitly
activated when building the library in order to use it.

Please note that this does not mean that the parser-iter feature will
never be the default feature; if it turns out that it is significantly
more performant than the manual implementation, then it stands a good
chance that the parser-iter feature will become the default
implementation. In all cases, however, it will always be the manual
implementation that defines the behaviour of path iteration.