Add recipes to justfile for feature specific code coverage
The justfile configuration only contains a recipe to run code coverage
for the default feature ie the manual-iter feature. However, it includes
files that are not run or even compiled -- files used only for the
parser-iter feature -- resulting in incorrect code coverage results.

This change adds recipes to the justfile configuration which handle:

- code coverage for the manual-iter feature

- code coverage for the parser-iter feature

- running unit tests for a specific feature

Please note that this change also updates variable assignment to use
`:=` instead of a plain `=`; without this change, the just command
issues warnings regarding the deprecation of `=` for assignment.