Update travisci config to run tests via the just command
The travisci configuration file includes commands to run unit tests and
code coverage. However, these commands are already represented in the
justfile configuration file. If how unit tests and/or code coverage are
run changes, then not only will the justfile configuration file need to
be updated, the travisci configuration file will also need to be
updated. This only serves to make it easier to make a mistake resulting
in either tests not being run, or builds outright failing.

This change removes the steps for running unit tests and code coverage,
and instead runs them via the just command.

Please note that the travisci configuration is now aware of the
project's features, and as such includes jobs to test each feature.