Update azure config to be aware of project features
The azure pipelines configuration only builds and runs tests for the
default features which is currently only the manual-iter feature. This
means that the parser-iter feature is not being validated.

This change updates the azure pipelines configuration so that it will
run tests for both the manual-iter feature and the parser-iter feature.