Add new benchmark: compare_iter
The relative performance of the manual-iter and parser-iter features are
unknown. It is believed that the manual-iter feature is the more
performant of the two features but this is not proven. The intention as
of this commit is for the manual-iter feature to be used as the default
path iterator implementation, with the parser-iter feature being an
alternative implementation that serves as a secondary validation of the
behaviour.

This change adds a new benchmark comparing the performance of the
manual-iter feature versus the parser-iter feature. A recipe to invoke
the benchmark is also added to the justfile configuration file.

Note: this benchmark disables the default benchmark harness in order to
use the harness provided provided by the criterion crate.