fix: unable to import from unix or windows iter modules
In order to create a benchmark comparing the performance of the
manual-iter and parser-iter features, benchmark code will need to import
necessary traits and structs from the unix::iter and windows::iter
modules. This, however, cannot be done as these modules are not public.

This change updates the following modules so that they are public:

- unix::iter

- windows::iter

- unix::iter::{manual, parser}

- windows::iter::{manual, parser}