Add new benchmark: compare_stdpath
The relative performance between the standard library's Path type and
the pathlib crate's UnixPath/WindowsPath types is unknown. Considering
the standard library's Path type is the baseline, it is unknown whether
further work needs to be done to at least match the baseline
performance.

This change adds a new benchmark comparing the performance of the
standard library's Path type against the pathlib crate's
UnixPath/WindowsPath type using the manual-iter feature. A recipe to
invoke the benchmark is also added to the justfile configuration file.

Note: this benchmark disables the default benchmark harness in order to
use the harness provided provided by the criterion crate.