Add new benchmark: compare_unix_win_iter
The relative performance between the unix path iter and the windows path
iter unknown. Thus, it is unknown whether further work needs to be done
for one of the platform path iter implementations so that the
performance characteristics are as similar as possible.

This change adds a new benchmark comparing the performance of the
UnixPath type's path iteration against the WindowsPath type's path
iteration.

Note: this benchmark disables the default benchmark harness in order to
use the harness provided provided by the criterion crate.