Run compare_stdpath benchmark on azure pipelines
The development of the pathlib library is done on the linux platform.
However, it is intended for this library to work everywhere. However,
while unit tests can detect if the library is not working properly on
the windows platform via continuous integration services, nothing is
setup to also check the performance of the library on windows.

This change adds an additional step to the azure pipelines configuration
that runs the compare_stdpath benchmark when on the windows platform.
The data generated by this can then give a general idea of how well the
pathlib library performs in comparison to the standard library Path
type.