Add new struct: ErrorInfo
Path iteration currently yields a Component variant wrapped in a Result.
This may induce an unnecessary overhead resulting in performance
slightly worse than iteration of the standard library's Path type.
While it may speed iteration up to no longer wrap Component types in a
Result, there then needs to be some other mechanism to identify a
parsing error. The idea, then, is for there to be an `Error` variant of
the Component enum that communicated that a parsing error occurred, and
once this happens for error information to be stored in the iterator.

One idea is to directly store a ParseError type. This is not so good,
however, as ParseError stores owned types such as OsString and String
which unnecessarily increases the size of the iterator. Instead, just
the information needed to create a ParseError should be stored.

This change adds a new struct `ErrorInfo` that contains strictly only
what is needed in order to construct a ParseError type. Importantly,
ErrorInfo stores only primitive types such as usize, or references such
as &str, helping keep memory usage down. However, a conversion from an
ErrorInfo to a full ParseError is implemented allowing future commits to
implement doing so on demand.