Save position of parsing error during path iteration
Some path component parsers are able to report the exact position where
a parsing error occurred. However, this information is not recorded in
either the ErrorInfo struct or the iterator struct itself making it not
possible to create more fine-grained error messages that communicates
the position of the error as opposed to just the start and end positions
of the path component that triggered the error.

This change adds the capability to store the position of a parsing error
that occurs during path iteration.