refactor: allow creating ParseError w/ custom message
The ErrorInfo struct can create a ParseError struct using the stored
string slice as the error's string message. However, this is not
flexible; for instance, a custom string message cannot be created using
the values stored in the ErrorInfo struct in a custom manner, the only
error message allowed currently is the string slice directly stored in
the ErrorInfo struct.

This change refactors ErrorInfo::to_error() so that the creation of the
ParseError struct allows for specifying a custom error message string.
As of this commit, specifying a custom error message is not yet used to
its full intended use, however future changes will provide a public
interface for providing a custom error message.