Add new ErrorInfo method: with_errmsg()
The ErrorInfo struct can create a ParseError struct using the stored
string slice as the error's string message. However, this is not
flexible; for instance, a custom string message cannot be created using
the values stored in the ErrorInfo struct in a custom manner, the only
error message allowed currently is the string slice directly stored in
the ErrorInfo struct.

This change adds a new method to the ErrorInfo struct `with_errmsg`.
This new method is a public interface for being able to use an ErrorInfo
object to construct a custom error string, and use that custom error
string to create a new ParseError object.

This new method will allow being able to create more meaningful error
messages suitable to be displayed to a user. For example, if a parse
error is detected when iterating over a path, the message stored will
normally not include any position information. However, by using the
ErrorInfo::with_errmsg method, the position information stored in the
ErrorInfo object can be used to create a new error message that includes
both the message and the position information. For a user, this can make
it much easier to figure out which byte in a path caused the parse error
to occur.