Change From<ErrorInfo> impl to From<&ErrorInfo>
When detecting that a parse error occurred, some times it is necessary
to have access to the ErrorInfo object after having converted it to a
ParseError object. For example, this is used in some unit tests in order
to validate that the correct error was generated. However, when path
iteration is stored in a collection (such as a Vec), the ErrorInfo
object that is generated is owned by the collection with only a
reference to the object being returned. This means the ErrorInfo object
cannot be moved out of the collection and, as a result, cannot be
converted into a ParseError object using the From implementation
provided in the crate::common::error module.

This change updates the From implementation so that it only converts a
reference to an ErrorInfo object. This allows for creating a new
ParseError object without consuming the ErrorInfo object, allowing for
access to ErrorInfo after conversion.