refactor: path iteration to yield bare Component enum variant
The path iterator types use a Component enum type wrapped in a Result as
the item yielded during iteration. However, doing this has a small
overhead resulting in demonstrably worse performance than the standard
library's Path type; for simple, short paths, the standard library Path
type can achieve around 260 nanoseconds to collect into a vector whereas
the pathlib path iteration does it in 300+ nanoseconds.

This commit changes the type yielded by iteration over a path to be a
bare Component enum. This results in performance on par and sometimes a
little better -- ie anywhere from 5ns to 20ns better -- than the
standard library's Path type for simple types.

Due to no longer relying on wrapping path components in a Result type in
order to detect errors -- the Component::Error variant is used for this
instead -- all tests are updated to reflect the changes made.