Record position of unix path iter error (manual-iter)
When a parsing error occurs, the exact position in the path that the
error occurred or starts at is not recorded; the most that can be
determined is the start position of the component that triggered the
error. This can make it difficult to determine what the exact issue is
when an error occurs especially when dealing with non-printable
characters (if viewing the path as a string) or bytes.

This change stored the exact byte position in the path that an error
starts at when iterating over a unix path, making this information
available when constructing an error message.

Please note that this change is only applicable when the manual-iter
feature is activated.