Identify windows path errors in iterator (manual-iter)
When a parsing error occurs, the exact position in the path that the
error occurred or starts at is not recorded; the most that can be
determined is the start position of the component that triggered the
error. This can make it difficult to determine what the exact issue is
when an error occurs especially when dealing with non-printable
characters (if viewing the path as a string) or bytes.

For the manual implementation of windows path iteration, errors are not
explicitly created or handled. Detection is done at the individual path
component token level such that if an error occurs, then the token is
simply evaluated as non-matching. Unfortunately, this throws away
information needed to be able to communicate the exact position where
an error occurred.

This change identifies errors in the parsing layer instead of in the
token layer; tokens now expect that data has no parsing errors, and can
instead focus on whether data matches a token or not. This makes it
easier to capture the position in the path where an error occurred.

Please note, that this change only moves error identification into the
parsing layer. Identifying the exact position in the path that a parsing
error occurred or starts at will be added in a future change.