Add accessor methods to ErrorInfo struct
The ErrorInfo struct is meant to hold the minimum amount of information
needed to construct a ParseError type without needing to allocate
memory. This information is private which is an issue when trying to
validate an ErrorInfo object, since the information cannot be accessed.
As a result, the ErrorInfo cannot be easily validated especially in unit
tests.

This change adds accessor methods to the ErrorInfo struct that allows
retrieving an immutable reference to the information stored in the
struct.