fix: unix parser fails to parse valid path component
When iterating over a unix path when using the parser-iter feature, and
the path contains a null byte, then the error that is generated reports
the wrong position of the error.

If parsing a component consumes input and then fails, then a parse error
is generated. However, since parsing a component uses sub-parsers to
detect the '..' and '.' components (ie parent directory and current
directory respectively), and those sub-parsers are used such that they
consume the input path bytes when failing, then the failure reported is
due to either failing to match '..' or '.' components. This then results
in the wrong position being reported as the start of the error.

This change fixes the issue by ensuring that sub-components do not
consume input if they fail to match.