Include error position info in windows path parse errors (manual-iter)
When a parsing error occurs while iterating over a windows path using
the manual-iter feature, the exact position of the error in the path is
not stored when a parse error is created. This means that any error
messages that are needed to be shown to the user some times may not have
enough information to know exactly where the issue is in the path.

This change updates the manual windows path iterator so that when a
parsing error occurs, the position in the path that the error starts at
is recorded and accessible when creating an error message from a parse
error. This allows for error messages that includes the exact position
in the path where the error occurs.