Replace use of valid_part_char() parser (parser-iter)
The valid_part_char() parser function is only used in a single place in
the entire codebase ie when converting the combine's ParseError type
into an error type native to the pathlib crate. As there is no need for
this function anywhere else, and since the behaviour provided by this
function is simple to replicate without needing to use a parser library,
the valid_part_char() function is thus considered more akin to cruft
that needs cleaning up.

For reference, the valid_part_char() parser function parses a byte
string checking if there are any invalid characters in the byte string.
This function is used to extract the next path component from a path.

This change replaces the use of the valid_part_char() with a simple
iteration over the byte string that manually checks each item in the
string and records the position in the string of a path separator.

Please note: this change only removes use of the valid_part_char()
function, the function itself is still defined in the crate. Removal of
the valid_part_char() function will be done in a future commit.