fix: windows path w/ device name being invalid (parser-iter)
With only the parser-iter feature activated, when iterating over a
windows path, non-prefix components that contain a device name somewhere
in the middle (ie doesn't start with the device name) is sometimes
incorrectly parsed as invalid; per windows filesystem path rules, it is
only when a component's name as a whole matches a device name (or in
other words the name starts with the device name and has nothing else
following besides a possible file extension) that the component should
be considered invalid. This is detected by some tests of the unit test
suite that randomly generates such inputs.

For example, the strings "COM1", "com1", and "com1.txt" should all be
considered invalid names, however the string " COM1" should be
considered valid. Currently, the string " COM1" is being flagged as
invalid which is incorrect.

The cause of this issue is due to some parsers using regular expressions
to check whether a windows non-prefix path component is valid. As none
of the regular expressions expressly state the pattern to match should
be at the beginning of the string, these parsers will flag a path
component as invalid if it matches the pattern found anywhere in a path
component's name.

This change ensures that parsers using regular expressions to validate
path components all match from the beginning of the component's name
instead of anywhere in the name.